scripts/requirements.txt
httpx
pydantic
## requires that ffmpeg be installed on the host
ffmpeg-python